Add docker-compose.yaml Rename grimoire.go
